Include a modified copy of the BSplineBasis class from dune-functions
This class already implements B-spline basis functions and their derivatives.
Hopefully it is helpful to get started with B-spline surfaces.
It is not a verbatim copy of the dune-functions file.  Instead, I ripped
out all references to other files in dune-functions, and deleted a few
methods that are not needed here.